.github: add "required" ci job (#276)
Github branch protection rules allow required checks to be specified, but they
must be mentioned by name. This is a little annoying for jobs with names that
change, such as test jobs that depend on the go version.

This PR fixes this by creating a stub job in the CI workflow that simply
depends on the test and lint workflows. This job can then be listed as a
required check, and has the effect of aggregating the dependent checks into
one.